Merge pull request #33 from Su5eD/unstable

+++ src/api/java/mods/gregtechmod/api/machine/IUpgradableMachine.java
@@ -68,7 +68,7 @@ public interface IUpgradableMachine extends IElectricMachine {
 
     double getStoredSteam();
 
-    double getSteamCapacity();
+    int getSteamCapacity();
 
     long getStoredMj();
